+++ apps/fluent-tester/src/E2E/FocusTrapZone/pages/FocusTrapZonePageObject.win.ts
@@ -3,14 +3,14 @@ import {
   FOCUSTRAPZONE_TEST_COMPONENT,
   HOMEPAGE_FOCUSTRAPZONE_BUTTON,
 } from '../../../TestComponents/FocusTrapZone/consts';
-import { BasePage, By } from '../../common/BasePage.win';
+import { BasePage, By } from '../../common/BasePage';
 
 class FocusTrapZonePageObject extends BasePage {
   // OVERRIDE: We use isExisting() here instead of isDisplayed() because FocusTrapZone does not have any UI to it, it's simply
   // a wrapper that adds keyboard focus functionality
-  waitForPrimaryElementDisplayed(timeout?: number): void {
+  async waitForPrimaryElementDisplayed(timeout?: number): Promise<void> {
     const errorMsg = 'The FocusTrapZone UI Element did not load correctly. Please see logs.';
-    this.waitForCondition(() => this._primaryComponent.isExisting(), errorMsg, timeout);
+    await this.waitForCondition(async () => await this._primaryComponent.isExisting(), errorMsg, timeout);
   }
 
   /*****************************************/