+++ apps/fluent-tester/src/E2E/FocusZone/specs/FocusZone.spec.macos.ts
@@ -1,23 +1,21 @@
-import NavigateAppPage from '../../common/NavigateAppPage.macos';
-import FocusZonePageObject from '../pages/FocusZonePageObject.macos';
+import NavigateAppPage from '../../common/NavigateAppPage';
+import FocusZonePageObject from '../pages/FocusZonePageObject';
 import { PAGE_TIMEOUT, BOOT_APP_TIMEOUT } from '../../common/consts';
 
 // Before testing begins, allow up to 60 seconds for app to open
 describe('FocusZone Testing Initialization', function () {
-  it('Wait for app load', () => {
-    NavigateAppPage.waitForPageDisplayed(BOOT_APP_TIMEOUT);
-    expect(NavigateAppPage.isPageLoaded()).toBeTruthy();
+  it('Wait for app load', async () => {
+    await NavigateAppPage.waitForPageDisplayed(BOOT_APP_TIMEOUT);
+    await expect(await NavigateAppPage.isPageLoaded()).toBeTruthy();
   });
 
-  it('Click and navigate to FocusZone test page', () => {
-    /* Scroll to component test page button in scrollview if not already visible*/
-    FocusZonePageObject.scrollToComponentButton();
-    FocusZonePageObject.waitForButtonDisplayed(PAGE_TIMEOUT);
+  it('Click and navigate to FocusZone test page', async () => {
+    await FocusZonePageObject.waitForButtonDisplayed(PAGE_TIMEOUT);
 
     /* Click on component button to navigate to test page */
-    NavigateAppPage.clickAndGoToFocusZonePage();
-    FocusZonePageObject.waitForPageDisplayed(PAGE_TIMEOUT);
+    await NavigateAppPage.clickAndGoToFocusZonePage();
+    await FocusZonePageObject.waitForPageDisplayed(PAGE_TIMEOUT);
 
-    expect(FocusZonePageObject.isPageLoaded()).toBeTruthy();
+    await expect(await FocusZonePageObject.isPageLoaded()).toBeTruthy();
   });
 });