+++ apps/fluent-tester/src/TestComponents/Avatar/CustomizedAvatar.tsx
@@ -1,12 +1,19 @@
 import React, { useState, useMemo } from 'react';
-import { Avatar, AvatarSize } from '@fluentui-react-native/avatar';
-import { View, Text, TextInput, Platform } from 'react-native';
+import { Avatar, AvatarColors, AvatarSize, getJavaHashCode } from '@fluentui-react-native/avatar';
+import { View, Text, TextInput, Platform, StyleSheet } from 'react-native';
 import { steveBallmerPhotoUrl } from './../PersonaCoin/styles';
 import { commonTestStyles as commonStyles } from '../Common/styles';
 import { FontWeight } from '@fluentui-react-native/theme-types';
 import { SvgIconProps } from '@fluentui-react-native/icon';
 import TestSvg from '../../FluentTester/test-data/test.svg';
 import { ToggleButton } from '@fluentui/react-native';
+
+const styles = StyleSheet.create({
+  avatarTestCaseContainer: { marginLeft: 20 },
+  avatarTokenOptions: { flexDirection: 'row' },
+  fontTokenOptions: { paddingHorizontal: 20 },
+  tokenHeader: { fontWeight: 'bold' },
+});
 
 export const CustomizeUsage: React.FunctionComponent = () => {
   const [showImage, setShowImage] = useState(true);
@@ -25,6 +32,7 @@ export const CustomizeUsage: React.FunctionComponent = () => {
   const [ringColor, setRingColor] = useState<string>(undefined);
   const [ringBackgroundColor, setRingBackgroundColor] = useState<string>('yellow');
   const [ringThickness, setRingThickness] = useState<string>('4');
+  const [ringInnerGap, setRingInnerGap] = useState<string>('4');
   const [showRing, setShowRing] = useState<boolean>(true);
 
   const CustomizedAvatar = useMemo(() => {
@@ -40,6 +48,7 @@ export const CustomizeUsage: React.FunctionComponent = () => {
       ringColor,
       ringBackgroundColor,
       ringThickness: parseInt(ringThickness),
+      ringInnerGap: parseInt(ringInnerGap),
     };
     return Avatar.customize(tokens);
   }, [
@@ -51,6 +60,7 @@ export const CustomizeUsage: React.FunctionComponent = () => {
     size,
     ringColor,
     ringBackgroundColor,
+    ringInnerGap,
     ringThickness,
     fontWeight,
     fontFamily,
@@ -61,6 +71,9 @@ export const CustomizeUsage: React.FunctionComponent = () => {
     src: TestSvg,
     viewBox: '0 0 500 500',
   };
+
+  const useJavaHashCode = ['ios', 'macos', 'android'].includes(Platform.OS);
+  const hashedColor = AvatarColors[getJavaHashCode(name) % (AvatarColors.length - 3)];
 
   return (
     <View style={commonStyles.root}>
@@ -74,7 +87,7 @@ export const CustomizeUsage: React.FunctionComponent = () => {
         <ToggleButton onClick={() => setShowRing(!showRing)} checked={showRing} style={commonStyles.vmargin}>
           {showRing ? 'Hide' : 'Show'} ring
         </ToggleButton>
-        <View style={{ flexDirection: 'row' }}>
+        <View style={styles.avatarTokenOptions}>
           <View>
             <TextInput
               accessibilityLabel="Name for generating initials"
@@ -94,7 +107,7 @@ export const CustomizeUsage: React.FunctionComponent = () => {
                 setInitials(e.nativeEvent.text);
               }}
             />
-            <Text style={{ fontWeight: 'bold' }}>Avatar tokens</Text>
+            <Text style={styles.tokenHeader}>Avatar tokens</Text>
             <TextInput
               accessibilityLabel="Background color"
               style={commonStyles.textBox}
@@ -133,7 +146,7 @@ export const CustomizeUsage: React.FunctionComponent = () => {
                 setIconColor(e.nativeEvent.text);
               }}
             />
-            <Text style={{ fontWeight: 'bold' }}>Ring tokens</Text>
+            <Text style={styles.tokenHeader}>Ring tokens</Text>
             <TextInput
               accessibilityLabel="Ring background color"
               style={commonStyles.textBox}
@@ -161,9 +174,18 @@ export const CustomizeUsage: React.FunctionComponent = () => {
                 setRingThickness(e.nativeEvent.text);
               }}
             />
+            <TextInput
+              accessibilityLabel="Ring inner gap"
+              style={commonStyles.textBox}
+              placeholder="Ring inner gap"
+              blurOnSubmit={true}
+              onSubmitEditing={(e) => {
+                setRingInnerGap(e.nativeEvent.text);
+              }}
+            />
           </View>
-          <View style={{ paddingHorizontal: 20 }}>
-            <Text style={{ fontWeight: 'bold' }}>Font tokens</Text>
+          <View style={styles.fontTokenOptions}>
+            <Text style={styles.tokenHeader}>Font tokens</Text>
             <TextInput
               accessibilityLabel="Initials text color"
               style={commonStyles.textBox}
@@ -217,7 +239,7 @@ export const CustomizeUsage: React.FunctionComponent = () => {
           transparentRing={!showRing}
         />
       </View>
-      <View style={{ marginLeft: 20 }}>
+      <View style={styles.avatarTestCaseContainer}>
         <Text>Avatar customized with props</Text>
         <Avatar
           active="active"
@@ -232,8 +254,27 @@ export const CustomizeUsage: React.FunctionComponent = () => {
           ringThickness={parseInt(ringThickness)}
           size={parseInt(size) as AvatarSize}
           transparentRing={!showRing}
+          ringInnerGap={parseInt(ringInnerGap)}
         />
       </View>
+      {useJavaHashCode && (
+        <View style={styles.avatarTestCaseContainer}>
+          <Text>Avatar with hashed color</Text>
+          <Avatar
+            active="active"
+            activeAppearance="ring"
+            avatarColor={hashedColor}
+            initials={showInitials ? initials : undefined}
+            name={showInitials ? name : undefined}
+            ringBackgroundColor={ringBackgroundColor}
+            ringColor={ringColor}
+            ringThickness={parseInt(ringThickness)}
+            size={parseInt(size) as AvatarSize}
+            transparentRing={!showRing}
+            ringInnerGap={parseInt(ringInnerGap)}
+          />
+        </View>
+      )}
     </View>
   );
 };