+++ packages/components/Notification/CHANGELOG.md
@@ -1,8 +1,111 @@
 # Change Log - @fluentui-react-native/notification
 
-This log was last generated on Fri, 05 Aug 2022 23:22:38 GMT and should not be manually modified.
+This log was last generated on Tue, 06 Sep 2022 18:22:03 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.17.3
+
+Tue, 06 Sep 2022 18:22:03 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/button to v0.24.20
+- Bump @fluentui-react-native/icon to v0.13.6
+- Bump @fluentui-react-native/interactive-hooks to v0.18.3
+- Bump @fluentui-react-native/pressable to v0.9.9
+- Bump @fluentui-react-native/text to v0.14.2
+
+## 0.17.2
+
+Wed, 31 Aug 2022 12:11:39 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/button to v0.24.19
+
+## 0.17.1
+
+Mon, 29 Aug 2022 23:21:44 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/button to v0.24.18
+- Bump @fluentui-react-native/experimental-shadow to v0.2.14
+- Bump @fluentui-react-native/framework to v0.8.1
+- Bump @fluentui-react-native/icon to v0.13.5
+- Bump @fluentui-react-native/interactive-hooks to v0.18.2
+- Bump @fluentui-react-native/pressable to v0.9.8
+- Bump @fluentui-react-native/text to v0.14.1
+
+## 0.17.0
+
+Fri, 26 Aug 2022 22:09:56 GMT
+
+### Minor changes
+
+- move spec to outer folder (joannaquu@gmail.com)
+
+## 0.16.1
+
+Fri, 26 Aug 2022 18:15:17 GMT
+
+### Patches
+
+- Use undefined shadowToken in NotificationTokens (78454019+lyzhan7@users.noreply.github.com)
+- Bump @fluentui-react-native/button to v0.24.17
+- Bump @fluentui-react-native/experimental-shadow to v0.2.13
+
+## 0.16.0
+
+Wed, 17 Aug 2022 20:36:34 GMT
+
+### Minor changes
+
+- update notification text font (joannaquu@gmail.com)
+
+## 0.15.2
+
+Mon, 15 Aug 2022 23:29:03 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/button to v0.24.16
+- Bump @fluentui-react-native/experimental-shadow to v0.2.12
+
+## 0.15.1
+
+Mon, 15 Aug 2022 22:12:03 GMT
+
+### Patches
+
+- Pass key shadow props into child instead of a new view (78454019+lyzhan7@users.noreply.github.com)
+- Bump @fluentui-react-native/button to v0.24.15
+- Bump @fluentui-react-native/experimental-shadow to v0.2.11
+
+## 0.15.0
+
+Mon, 15 Aug 2022 18:42:24 GMT
+
+### Minor changes
+
+- remove minHeight: 62 (joannaquu@gmail.com)
+
+## 0.14.0
+
+Fri, 12 Aug 2022 23:18:50 GMT
+
+### Minor changes
+
+- update minHeight (joannaquu@gmail.com)
+
+## 0.13.0
+
+Wed, 10 Aug 2022 22:24:35 GMT
+
+### Minor changes
+
+- reorder notification voiceover (joannaquu@gmail.com)
 
 ## 0.12.0