+++ packages/experimental/Switch/CHANGELOG.md
@@ -1,8 +1,67 @@
 # Change Log - @fluentui-react-native/switch
 
-This log was last generated on Mon, 08 Aug 2022 21:40:41 GMT and should not be manually modified.
+This log was last generated on Tue, 06 Sep 2022 18:22:03 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.6.4
+
+Tue, 06 Sep 2022 18:22:03 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/text to v0.14.2
+- Bump @fluentui-react-native/interactive-hooks to v0.18.3
+
+## 0.6.3
+
+Mon, 29 Aug 2022 23:21:44 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/text to v0.14.1
+- Bump @fluentui-react-native/framework to v0.8.1
+- Bump @fluentui-react-native/interactive-hooks to v0.18.2
+
+## 0.6.2
+
+Fri, 26 Aug 2022 18:15:17 GMT
+
+### Patches
+
+- Added better HC token mapping for win32 (nkhalil942@gmail.com)
+
+## 0.6.1
+
+Fri, 19 Aug 2022 15:32:12 GMT
+
+### Patches
+
+- Added more space between the On/Off text, Label, and Track to align more with Fluent v9. (nkhalil942@gmail.com)
+
+## 0.6.0
+
+Wed, 17 Aug 2022 18:45:52 GMT
+
+### Minor changes
+
+-  Label position set to 'after' will prevent onText and offText from rendering (nkhalil942@gmail.com)
+
+## 0.5.0
+
+Thu, 11 Aug 2022 22:49:12 GMT
+
+### Minor changes
+
+- Replaced LayoutAnimation with animationClass (duomaster01@gmail.com)
+
+## 0.4.1
+
+Wed, 10 Aug 2022 18:25:08 GMT
+
+### Patches
+
+- Added Switch Documentation (nkhalil942@gmail.com)
 
 ## 0.4.0