+++ packages/theming/win32-theme/CHANGELOG.md
@@ -1,8 +1,18 @@
 # Change Log - @fluentui-react-native/win32-theme
 
-This log was last generated on Thu, 28 Jul 2022 23:57:03 GMT and should not be manually modified.
+This log was last generated on Mon, 29 Aug 2022 23:21:44 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.21.0
+
+Mon, 29 Aug 2022 23:21:44 GMT
+
+### Minor changes
+
+- Added unit tests (nkhalil942@gmail.com)
+- Bump @fluentui-react-native/default-theme to v0.15.0
+- Bump @fluentui-react-native/theming-utils to v0.15.0
 
 ## 0.20.1